Added loading message before the scenes are loaded to let the user know what's happening.
Fixed the Info window so that it can actuall display things.

Implemented a per-scene information readout.

TODO: fill out the information which is to be spit out per scene in order to more fully fit the requrements set out by the professor, such as describing high salary and showing that most companies have food courts, etc.